requirements.txt
affine==2.4.0
aim==3.23.0
aim-ui==3.23.0
aimrecords==0.0.7
aimrocks==0.5.2
aiofiles==24.1.0
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
albucore==0.0.13
albumentations==1.4.13
alembic==1.13.2
annotated-types==0.7.0
anyio==4.4.0
attrs==24.1.0
base58==2.0.1
boto3==1.34.154
botocore==1.34.154
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.2.1
cryptography==43.0.0
cycler==0.12.1
eval_type_backport==0.2.0
fastapi==0.112.0
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
geopandas==1.0.1
greenlet==3.0.3
h11==0.14.0
h5py==3.11.0
huggingface-hub==0.24.5
idna==3.7
imageio==2.34.2
iniconfig==2.0.0
Jinja2==3.1.4
jmespath==1.0.1
kiwisolver==1.4.5
lazy_loader==0.4
lightning-utilities==0.11.6
Mako==1.3.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.3
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pluggy==1.5.0
psutil==6.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pyogrio==0.9.0
pyparsing==3.1.2
pyproj==3.6.1
pytest==8.3.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.3.3
pytz==2024.1
PyYAML==6.0.1
rasterio==1.3.10
regex==2024.7.24
requests==2.32.3
RestrictedPython==7.2
s3transfer==0.10.2
safetensors==0.4.4
scikit-image==0.24.0
scipy==1.14.0
shapely==2.0.5
six==1.16.0
sniffio==1.3.1
snuggs==1.4.7
SQLAlchemy==2.0.32
starlette==0.37.2
sympy==1.13.1
tifffile==2024.7.24
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.0
torchmetrics==1.4.1
torchvision==0.19.0
tqdm==4.66.5
transformers==4.43.4
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.5
websockets==12.0
yarl==1.9.4
